Renaming the file customError.ts to custom-error.ts
